+++ R/document.r
@@ -11,11 +11,12 @@
 #'   introspection on the code objects to determine how to document them.
 #' @keywords programming
 #' @export
-#' @importFrom digest digest
 document <- function(pkg = ".", clean = FALSE,
   roclets = c("collate", "namespace", "rd"), reload = TRUE) {
+  if (is_installed("roxygen2", 3)) {
+    stop("Please install latest roxygen2", call. = FALSE)
+  }
 
-  require("roxygen2")
   pkg <- as.package(pkg)
   message("Updating ", pkg$package, " documentation")
 
@@ -23,48 +24,21 @@ document <- function(pkg = ".", clean = FALSE,
   if (!file.exists(man_path)) dir.create(man_path)
 
   if (clean) {
-    if (packageVersion("roxygen2") < 3) {
-      roxygen2::clear_caches()  
-    }
     file.remove(dir(man_path, full.names = TRUE))
   }
-  
+
   if (!is_loaded(pkg) || (is_loaded(pkg) && reload)) {
     try(load_all(pkg, reset = clean))
   }
 
-  if (packageVersion("roxygen2") < 3) {
-    document_roxygen2(pkg, roclets)
-  } else {
-    document_roxygen3(pkg, roclets)
-  }
-  
+  document_roxygen3(pkg, roclets)
+
   clear_topic_index(pkg)
   invisible()
 }
 
-document_roxygen2 <- function(pkg, roclets) {  
-  # Integrate source and evaluated code
-  env <- ns_env(pkg)
-  env_hash <- suppressWarnings(digest(env))
-  r_files <- find_code(pkg)
-  parsed <- unlist(lapply(r_files, parse.file, env = env,
-    env_hash = env_hash), recursive = FALSE)
-  
-  roclets <- paste(roclets, "_roclet", sep = "")
-  for (roclet in roclets) {
-    roc <- match.fun(roclet)()
-    with_envvar(r_env_vars(),
-      with_collate("C", {
-        results <- roxygen2:::roc_process(roc, parsed, pkg$path)
-        roxygen2:::roc_output(roc, results, pkg$path)
-      })
-    )
-  }
-}
-
-document_roxygen3 <- function(pkg, roclets) {  
-  with_envvar(r_env_vars(), with_collate("C", 
-    roxygenise(pkg$path, roclets = roclets, load_code = pkg_env)
+document_roxygen3 <- function(pkg, roclets) {
+  with_envvar(r_env_vars(), with_collate("C",
+    roxygen2::roxygenise(pkg$path, roclets = roclets, load_code = pkg_env)
   ))
 }