+++ NEWS.md
@@ -1,5 +1,7 @@
 # devtools 1.4.1.99
 
+* `document()` now requires at least version 3.0.0 of roxygen2.
+
 * `install_github()` now defaults to `dependencies = TRUE` so you definitely
   get all the packages you need to build from source.